fix(scripts): use greadlink from coreutils on macOS

+++ scripts/publish.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-DIR=$(dirname $(dirname $(readlink -f $0)))
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  READLINK=greadlink
+else
+  READLINK=readlink
+fi
+DIR=$(dirname $(dirname $($READLINK -f $0)))
+if [ -z $DIR ]; then exit 1; fi;
 rm -fr $DIR/dist
 git clone --branch=gh-pages --depth=1 git@github.com:ipy/test.git $DIR/dist
 rm -fr $DIR/dist/*
